+++ samples/local-ts-consumer/kt/src/jsMain/kotlin/index.kt
@@ -1,0 +1,10 @@
+@file:OptIn(ExperimentalJsExport::class)
+
+package dev.petuska.samples.ts
+
+fun greeting(name: String): String = "Hello $name"
+
+@JsExport
+fun greet(name: String) {
+  println(greeting(name))
+}