+++ samples/local-ts-consumer/ts/src/index.ts
@@ -1,0 +1,3 @@
+import { dev } from "@mpetuska/kt";
+
+dev.petuska.samples.ts.greet("General Kenobi")